% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_alphabeta}
\alias{calculate_alphabeta}
\title{Calculate parameters of beta distribution}
\usage{
calculate_alphabeta(mu, var)
}
\arguments{
\item{mu}{Numeric value for mean.}

\item{var}{Numeric value for variance}
}
\description{
Calculate alpha beta values of beta distribution from mean and variance.
}
