% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_beta}
\alias{mean_beta}
\title{Calculate mean from alpha/beta parameters of beta distribution}
\usage{
mean_beta(alpha, beta)
}
\arguments{
\item{alpha}{Numeric value.}

\item{beta}{Numeric value.}
}
\description{
Calculate mean from alpha/beta parameters of beta distribution
}
