% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_keep_order}
\alias{merge_keep_order}
\title{Base R merge with order maintained}
\usage{
merge_keep_order(data1, data2, idrow = NULL, sort = FALSE, ...)
}
\arguments{
\item{data1}{Input data 1.}

\item{data2}{Input data 2.}

\item{idrow}{Column that contains rows of IDs}

\item{sort}{Boolean to determine sorting.}
}
\description{
Can be used with R code to ensure adj matrix order is maintained.
}
