% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rule_check}
\alias{rule_check}
\title{Check if contact probabilities have violations to rule set}
\usage{
rule_check(var_list, rule_list, as_matrix = TRUE)
}
\arguments{
\item{var_list}{Columns containing values of contact patterns combinations (e.g. gender and age).}

\item{rule_list}{List of rules.}

\item{as_matrix}{Boolean.}
}
\description{
Compare list of rules againt contact matrix probabilities to help assign violations to 0.
}
