% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ruletest}
\alias{ruletest}
\title{Create validity matrix for contact patterns}
\usage{
ruletest(data, rule_set)
}
\arguments{
\item{data}{Dataset containing the combinations of contact patterns.}

\item{rule_set}{List of rules.}
}
\description{
Used in creating contact patterns with restrictions.
}
