% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_traits.R
\docType{class}
\name{S4BasicTrait}
\alias{S4BasicTrait}
\title{S4 class for basic population traits}
\description{
Class for containing basic demographic traits.
}
\section{Slots}{

\describe{
\item{\code{value}}{Input value for a particular trait of the population.}

\item{\code{range}}{Range of possible values for a particular trait of the population.}
}}

\seealso{
\code{\link{S4Population}}
}
