% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_traits.R
\docType{class}
\name{S4ContactTrait}
\alias{S4ContactTrait}
\title{S4 class for population contact traits}
\description{
Class for containing contact trait details
}
\section{Slots}{

\describe{
\item{\code{value}}{Input value for a particular trait of the population.}

\item{\code{range}}{Range of possible values for a particular trait of the population.}

\item{\code{contact_matrix}}{Contact probabilities between defined types of individuals in population.}

\item{\code{adj_matrix}}{Sparse adjacency matrix for population.}
}}

\seealso{
\code{\link{S4Population}}
}
