% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_disease.R
\docType{class}
\name{S4Disease}
\alias{S4Disease}
\alias{create_disease}
\title{S4 class for disease}
\usage{
create_disease(inf_prob, inf_p)
}
\arguments{
\item{inf_prob}{Integer value for probability of infection.}

\item{inf_p}{Integer value for infectious period.}
}
\description{
Class for containing information on disease to simulation across a population.
}
\section{Functions}{
\itemize{
\item \code{create_disease}: Constructor for Disease class
}}

\section{Slots}{

\describe{
\item{\code{inf_prob}}{Integer value for probability of infection; used in calculation of \code{beta (new infected) = inf_prob * # of successful contacts}.}

\item{\code{infectious_period}}{Integer value for infectious period.}

\item{\code{latent_period}}{Integer value for length of latent period (NOT IN USE).}

\item{\code{incubation_period}}{Integer value for length of incubation period (NOT IN USE).}

\item{\code{convalescence_period}}{Integer value for length of convalescence period; defines minimum time until can be infected again, otherwise known as the immunity period (NOT IN USE).}

\item{\code{severity_coef}}{Integer value for severity factor of disease (NOT IN USE).}
}}

\seealso{
\code{\link{S4Population}}
}
