% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_fit.R
\docType{class}
\name{S4Fit}
\alias{S4Fit}
\title{S4 class for fitting of disease simulation}
\description{
Class for fitting the simulations of disease processes
}
\section{Slots}{

\describe{
\item{\code{timesteps}}{Numeric value.}
}}

