% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_population.R
\docType{class}
\name{S4Population}
\alias{S4Population}
\alias{create_population}
\alias{show,Population-method}
\alias{get_age,Population-method}
\alias{set_age,Population-method}
\alias{get_gender,Population-method}
\alias{set_gender,Population-method}
\alias{set_contacts,Population-method}
\title{S4 class for population}
\usage{
create_population(n)

\S4method{show}{Population}(object)

\S4method{get_age}{Population}(pop_obj, value = NULL)

\S4method{set_age}{Population}(pop_obj, value = NULL, range)

\S4method{get_gender}{Population}(pop_obj, value = NULL)

\S4method{set_gender}{Population}(pop_obj, value = NULL, range)

\S4method{set_contacts}{Population}(
  pop_obj,
  value = NULL,
  range,
  vars = c(),
  rule_list = NULL,
  mu,
  variance,
  random_fill = TRUE,
  progress = TRUE
)
}
\arguments{
\item{n}{Integer, size of population.}

\item{object}{S4Population class.}

\item{pop_obj}{S4Population class.}

\item{value}{Vector of user-provided values for set operation (default: NULL, assigns a sample based on \code{range} parameter).}

\item{range}{Vector of min/max values for valid age.}

\item{vars}{Vector of variable names of interest to create contact patterns (e.g. gender and age).}

\item{rule_list}{List of values defining valid contact patterns (default: NULL, all combinations of \code{vars} considered probable).}

\item{mu}{Numeric value for \eqn{mu} for beta distribution (\eqn{mu = a / (a + b)}); determines heterogeneity of contact patterns.}

\item{variance}{Numeric value for beta distribution; used with \eqn{mu} to determine beta distribution shape parameters for contact pattern sampling.}

\item{random_fill}{Boolean value to determine if population is filled in random ordering (default: \code{TRUE}); otherwise will go in provided row order.}

\item{progress}{Boolean value to determine whether or not a progress bar is shown for creating the network; helfpul if you need networks for population of 10,000 and over.}
}
\description{
An S4 class to create a population for disease simulation.
}
\section{Functions}{
\itemize{
\item \code{create_population}: Constructor for Population class

\item \code{show,Population-method}: Show method for Population class

\item \code{get_age,Population-method}: Method for retrieving age structure of Population class

\item \code{set_age,Population-method}: Method for setting age structure of Population class

\item \code{get_gender,Population-method}: Method for retrieving gender structure of Population class

\item \code{set_gender,Population-method}: Method for setting gender structure of Population class

\item \code{set_contacts,Population-method}: Method for setting contact structure of Population class
}}

\section{Slots}{

\describe{
\item{\code{n}}{Integer, size of population.}

\item{\code{age_structure}}{S4 class of 'BasicTrait' depicting age details of population.}

\item{\code{gender_structure}}{S4 class of 'BasicTrait' depicting gender details of population.}

\item{\code{contact_structure}}{S4 class of 'ContactTrait' depicting contact structure of population.}
}}

