% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_simulate.R
\docType{class}
\name{S4Simulate}
\alias{S4Simulate}
\alias{run_simulation}
\title{S4 class for disease simulation}
\usage{
run_simulation(
  timesteps,
  init_inf,
  replications = 1,
  Population,
  Disease,
  random_init = TRUE,
  transition_model
)
}
\arguments{
\item{timesteps}{Numeric value.}

\item{init_inf}{Numeric value.}

\item{replications}{Integer value (NOT INTEGERATED YET...PERHAPS METHOD INSTEAD?).}

\item{Population}{S4Population class.}

\item{Disease}{S4Disease class.}

\item{random_init}{Boolean to control if random set of infected in population selected for each simulation.}

\item{transition_model}{Character or function to describe how transition occurs between disease states. Defaults include: SIS, SIR, and SEIR.}
}
\description{
Class for simulating disease process using population and disease classes

Creates the simulation class based upon provided input classes.
}
\section{Functions}{
\itemize{
\item \code{run_simulation}: Constructor for disease simulation class
}}

\section{Slots}{

\describe{
\item{\code{timesteps}}{Numeric value.}

\item{\code{Population}}{S4Population class.}

\item{\code{Disease}}{S4Disease class.}
}}

