% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_beta}
\alias{var_beta}
\title{Calculate variance from alpha/beta parameters of beta distribution}
\usage{
var_beta(alpha, beta)
}
\arguments{
\item{alpha}{Numeric value.}

\item{beta}{Numeric value.}
}
\description{
Calculate variance from alpha/beta parameters of beta distribution
}
